+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.3.3] - 2025-04-17
+
+### Scrizer Colors
+
+- fix Colors Themes
+- fix Colors extnsion
+
 ## [1.3.2] - 2025-04-16
 
 ### Scrizer Pubspec file